Modified COM interfaces (IUnknown and others) to use UINT32 instead of ULONG. On 32-bit ULONG is 32 bits long in C++; this gets translated to UInt32 in .NET. However, on 64-bit Linux machines ULONG is 64 bits long, therefore we have to use UINT32 in all COM interfaces that are used in managed and unmanaged code.
Reviewed by: Tom Hindle



svn-revision: 457

+++ Win32More/include/ErrorObjects.h
@@ -58,8 +58,8 @@ public:
 
 	// IUknown Methods
         virtual HRESULT STDMETHODCALLTYPE QueryInterface(REFIID riid, void ** ppv);
-        virtual ULONG   STDMETHODCALLTYPE AddRef();
-        virtual ULONG   STDMETHODCALLTYPE Release();
+        virtual UINT32   STDMETHODCALLTYPE AddRef();
+        virtual UINT32   STDMETHODCALLTYPE Release();
 
 	
 protected:
